Add test for remove task functionality
