+++ package.json
@@ -4,7 +4,8 @@
   "description": "![](https://img.shields.io/badge/Microverse-blueviolet)",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "watch": "jest --watch *.js",
     "build": "webpack",
     "start": "webpack serve --open"
   },
@@ -20,6 +21,7 @@
   },
   "homepage": "https://github.com/Mike-devloper/To-Do-list#readme",
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.1",
     "eslint": "^7.32.0",
@@ -27,6 +29,8 @@
     "eslint-plugin-import": "^2.26.0",
     "hint": "^7.1.2",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.1.2",
+    "jest-environment-jsdom": "^29.1.2",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
