+++ src/module/unitTest/deleteFunction.js
@@ -1,0 +1,5 @@
+const removeTask = (taskArray, ptask) => {
+  taskArray = taskArray.filter((task) => task.description !== ptask.description);
+}
+
+export default removeTask;
